feat: moved proxy structure to it's header file
